sam3env_requirements-gpu5070.txt
# --- Custom Index for RTX 50-series (Blackwell) Support ---
--extra-index-url https://download.pytorch.org/whl/nightly/cu128

# --- Core Blackwell Dependencies ---
torch>=2.7.0.dev
torchvision>=0.22.0.dev
torchaudio>=2.7.0.dev

# --- Existing Project Dependencies ---
addict==2.4.0
aiofiles==24.1.0
aiohttp==3.13.2
anyio==4.12.0
argon2-cffi==25.1.0
arrow==1.3.0
asttokens==2.2.1
attrs==23.1.0
beautifulsoup4==4.14.3
black==24.2.0
bleach==6.2.0
boto3==1.40.64
botocore==1.40.64
certifi==2025.11.12
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.0
click==8.3.1
comm==0.2.3
contourpy==1.3.2
coverage==7.12.0
cryptography==42.0.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.35.2
einops==0.8.1
exceptiongroup==1.3.1
executing==1.2.0
fairscale==0.4.13
fastapi==0.119.0
fastjsonschema==2.21.2
ffmpeg-python==0.2.0
filelock==3.20.0
fsspec==2025.12.0
ftfy==6.1.1
gitdb==4.0.12
GitPython==3.1.31
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.2.3
idna==3.11
imageio==2.37.0
iopath==0.1.10
ipykernel==6.30.1
ipython==8.16.1
ipywidgets==8.1.8
jedi==0.19.0
Jinja2==3.1.6
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.3
Pillow==11.3.0
portalocker==3.2.0
pycocotools==2.0.10
pyparsing==3.2.5
python-dateutil==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.0
rich==14.2.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
safetensors==0.7.0
timm==1.0.22
tqdm==4.67.1
typing-extensions==4.15.0
ultralytics==8.3.225
webcolors==24.11.1
yarl==1.22.0